requirements.txt
google-adk>=0.1.0
fastapi>=0.110.0
python-dotenv>=1.0.0
pydantic>=2.0.0
